third_party/WebKit/LayoutTests/inspector/sources/debugger/properties-special.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/console-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script>

function onload()
{
    console.dir(Object(true));
    console.dir((function(a,b) { return a + b; }));
    console.dir((function(a,b) { return a + b; }).bind({}, 2));
    console.dir((function*() { yeild [1,2,3] }));
    runTest();
}

function test()
{
    InspectorTest.expandConsoleMessages(dumpConsoleMessages);

    function dumpConsoleMessages()
    {
        InspectorTest.dumpConsoleMessages(false, false, InspectorTest.textContentWithLineBreaks);
        InspectorTest.completeTest();
    }
}

</script>
</head>
<body onload="onload()">
<p>Tests how debugger presents special properties of closures, bound functions and object wrappers.</p>
</body>
</html>
