third_party/WebKit/LayoutTests/inspector/sources/debugger/properties-special-expected.txt

CONSOLE MESSAGE: line 10: true
CONSOLE MESSAGE: line 11: function (a,b) { return a + b; }
CONSOLE MESSAGE: line 12: function () { [native code] }
Tests how debugger presents special properties of closures, bound functions and object wrappers.

properties-special.html:10 Boolean
    __proto__: Boolean
    [[PrimitiveValue]]: true
properties-special.html:11 anonymous(a, b)
    arguments: null
    caller: null
    length: 2
    name: ""
    prototype: Object
    __proto__: ()
    <function scope>
properties-special.html:12 anonymous()
    arguments: (...)
    caller: (...)
    length: 1
    name: "bound "
    __proto__: ()
    [[TargetFunction]]: (a,b)
    [[BoundThis]]: Object
    [[BoundArgs]]: Array[1]
